% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tidy_clin.R
\docType{data}
\name{data_tidy_clin_infant}
\alias{data_tidy_clin_infant}
\alias{data_tidy_clin_infant,data_tidy_clin_adult,data_tidy_clin,clinical_data,data_clinical}
\alias{data_tidy_clin_adult}
\title{Clinical data}
\format{
Dataframes with the following variables
\describe{
  \item{pid}{character. Participant ID.}
  \item{age}{"infant" or "adult". Age category of participant.}
  \item{race}{"black" or "non_black". Race of participant.
  Vast majority of non-black participants
  are mixed race.}
  \item{sex}{"male" or "female". Sex of participant.}
  \item{bcg}{"bcg" or "no bcg". BCG status. Infants only.}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 4 columns.
}
\usage{
data_tidy_clin_infant

data_tidy_clin_adult
}
\description{
Datasets containing clinical data for infants (\code{data_tidy_clin_infant})
and adults (\code{data_tidy_clin_adult}). The infant dataset
has one extra variable, \code{bcg}, indicating BCG status.
}
\keyword{datasets}
