% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tidy_freq_ifng.R
\docType{data}
\name{data_tidy_freq_ifng}
\alias{data_tidy_freq_ifng}
\title{Frequencies of IFNg+ cells}
\format{

}
\usage{
data_tidy_freq_ifng
}
\description{
Background-subtracted frequencies of
IFNg+ cells.
}
\keyword{datasets}
